requirements-dev.txt
# Development and Testing Dependencies
# Install with: pip install -r requirements-dev.txt

# Testing frameworks
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code quality tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Type checking
mypy>=1.5.0

# HTTP testing
httpx>=0.24.0
requests-mock>=1.11.0

# Environment and configuration testing
python-dotenv>=1.0.0

# Include main dependencies
-r backend/requirements.txt
